test: consolidate model catalog lookup coverage
